@@ -1,23 +1,15 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
+import { TestBed } from '@angular/core/testing';
 import { ReferentielListComponent } from './referentiel.list.component';
+import { AppComponent } from "../app.component";
 
 describe('ReferentielListComponent', () => {
-  let component: ReferentielListComponent;
-  let fixture: ComponentFixture<ReferentielListComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
+  beforeEach(() => TestBed.configureTestingModule({
       declarations: [ ReferentielListComponent ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(ReferentielListComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+  }));
 
   it('should create', () => {
-    expect(component).toBeTruthy();
+    const fixture = TestBed.createComponent(AppComponent);
+    const app = fixture.componentInstance;
+    expect(app).toBeTruthy();
   });
 });
